recompose files
