+++ js/packages/openinference-tanstack-ai/src/constants.ts
@@ -1,0 +1,4 @@
+export const INSTRUMENTATION_NAME = "@arizeai/openinference-tanstack-ai";
+export const AGENT_SPAN_NAME = "TanStack AI chat";
+export const TOOL_SPAN_PREFIX = "TanStack AI tool: ";
+export const LLM_SPAN_PREFIX = "TanStack AI LLM ";
